requirements.txt
pyyaml
openapi-spec-validator
requests
pyOpenSSL
